+++ Tests/ReactiveSwiftTests/SignalSpec.swift
@@ -678,10 +678,10 @@ class SignalSpec: QuickSpec {
 			}
 		}
 
-		describe("filterMap") {
+		describe("compactMap") {
 			it("should omit values from the signal that are nil after the transformation") {
 				let (signal, observer) = Signal<String, Never>.pipe()
-				let mappedSignal: Signal<Int, Never> = signal.filterMap { Int.init($0) }
+				let mappedSignal: Signal<Int, Never> = signal.compactMap(Int.init)
 
 				var lastValue: Int?
 
@@ -701,7 +701,7 @@ class SignalSpec: QuickSpec {
 
 			it("should stop emiting values after an error") {
 				let (signal, observer) = Signal<String, TestError>.pipe()
-				let mappedSignal: Signal<Int, TestError> = signal.filterMap { Int.init($0) }
+				let mappedSignal: Signal<Int, TestError> = signal.compactMap(Int.init)
 
 				var lastValue: Int?
 
@@ -724,7 +724,7 @@ class SignalSpec: QuickSpec {
 
 			it("should stop emiting values after a complete") {
 				let (signal, observer) = Signal<String, Never>.pipe()
-				let mappedSignal: Signal<Int, Never> = signal.filterMap { Int.init($0) }
+				let mappedSignal: Signal<Int, Never> = signal.compactMap(Int.init)
 
 				var lastValue: Int?
 
@@ -743,7 +743,7 @@ class SignalSpec: QuickSpec {
 
 			it("should send completed") {
 				let (signal, observer) = Signal<String, Never>.pipe()
-				let mappedSignal: Signal<Int, Never> = signal.filterMap { Int.init($0) }
+				let mappedSignal: Signal<Int, Never> = signal.compactMap(Int.init)
 
 				var completed: Bool = false
 
@@ -755,7 +755,7 @@ class SignalSpec: QuickSpec {
 
 			it("should send failure") {
 				let (signal, observer) = Signal<String, TestError>.pipe()
-				let mappedSignal: Signal<Int, TestError> = signal.filterMap { Int.init($0) }
+				let mappedSignal: Signal<Int, TestError> = signal.compactMap(Int.init)
 
 				var failure: TestError?